% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_simple.R
\name{poster}
\alias{poster}
\title{Posterdown HTML format (using pagedown::poster_relaxed)}
\usage{
poster(
  ...,
  template = pkg_resource("html", "template_poster.html"),
  css = NULL
)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::html_document}}

\item{template}{See \code{\link[pagedown]{html_paged}()}.}

\item{css}{See \code{\link[pagedown]{html_paged}()}.}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
Posterdown HTML format (using pagedown::poster_relaxed)
}
\examples{
\donttest{
file <- file.path(tempdir(),"foo.rmd")
rmarkdown::draft(file, template="poster", package="ovpaged")
}

}
