% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_simple.R
\name{simple}
\alias{simple}
\alias{simple_landscape}
\title{Simple paged report template}
\usage{
simple(css = NULL, autolink_bare_uris = FALSE, ...)

simple_landscape(css = NULL, autolink_bare_uris = FALSE, ...)
}
\arguments{
\item{css}{string: additional css}

\item{autolink_bare_uris}{logical: if \code{TRUE}, convert absolute URIs into links, even when not surrounded by pointy braces \verb{<...>}. It is set to \code{FALSE} by default here so that the \code{header-image} parameter in the document YAML is not converted. This is equivalent to specifying \code{md_extensions: -autolink_bare_uris} in the document YAML directly}

\item{...}{: passed to \code{\link[pagedown:html_paged]{pagedown::html_paged()}}}
}
\value{
A pagedown template
}
\description{
Simple paged report template
}
