% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_simple.R
\name{poster}
\alias{poster}
\title{Posterdown HTML format (using pagedown::poster_relaxed)}
\usage{
poster(
  ...,
  template = pkg_resource("html", "template_poster.html"),
  css = NULL
)
}
\arguments{
\item{...}{Additional arguments to \code{\link[pagedown:poster_relaxed]{pagedown::poster_relaxed()}}}

\item{template}{string: as for \code{\link[pagedown:html_paged]{pagedown::html_paged()}}}

\item{css}{string: additional css}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
Posterdown HTML format (using pagedown::poster_relaxed)
}
\examples{
 \donttest{
 myfile <- file.path(tempdir(), "foo.rmd")
 rmarkdown::draft(myfile, template = "poster", package = "ovpaged")
}

}
