% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_simple.R
\name{simple}
\alias{simple}
\alias{simple_landscape}
\title{Simple paged report template}
\usage{
simple(css = NULL, ...)

simple_landscape(css = NULL, ...)
}
\arguments{
\item{css}{string: additional css}

\item{...}{: passed to \code{\link[pagedown:html_paged]{pagedown::html_paged()}}}
}
\value{
A pagedown template
}
\description{
Simple paged report template
}
